requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
psycopg==3.2.3
psycopg-binary==3.2.3
tzdata==2024.2
wheel==0.44.0
pillow==11.0.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0